Renaming correct path
