+++ app/src/main/java/com/example/todotracker/ui/overdue/OverdueFragment.kt
@@ -1,12 +1,10 @@
-package com.example.todotracker.ui.queue
+package com.example.todotracker.ui.overdue
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import com.example.todotracker.R
 
@@ -21,7 +19,7 @@ class OverdueFragment : Fragment() {
     ): View? {
         queueViewModel =
                 ViewModelProvider(this).get(OverdueViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_queue, container, false)
+        val root = inflater.inflate(R.layout.fragment_overdue, container, false)
 //        val textView: TextView = root.findViewById(R.id.text)
 //        queueViewModel.text.observe(viewLifecycleOwner, Observer {
 //            textView.text = it
