+++ app/src/main/java/com/example/todotracker/ui/overdue/OverdueViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.todotracker.ui.queue
+package com.example.todotracker.ui.overdue
 
 import androidx.lifecycle.*
 import com.example.todotracker.model.ToDo
