+++ app/src/main/res/layout/fragment_overdue.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.queue.OverdueFragment">
+    tools:context=".ui.overdue.OverdueFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_queue"
